DDF-2344 Pull out functionality from CatalogFrameworkImpl to logical subcomponents.
This is the first phase of a massive restructuring/refactoring of CatalogFrameworkImpl.
